feat: implement hooks reducers for managing request and collection hooks scripts
